프레임워크 작성중 tcp 클래스 작성중
