+++ WINProgramingProject-master/WINProgramingProject-master/NGP_Server/NGP_Server/stdafx.cpp
@@ -1,25 +1,2 @@
 #include "stdafx.h"
 
-
-EROUND
-
-typedef struct udpdata
-{
-	char type;
-	char playerID;
-	int playerXpos;
-	int playerYpos;
-	int collisionDamage;
-	EROUND sceneState;
-}udpData;
-
-typedef struct tcpdata
-{
-	char type;
-	char playerID;
-	bool useTeleport;
-	int teleportXpos;
-	int teleportYpos;
-	bool useDash;
-	int hp;
-}tcpData;
