+++ WINProgramingProject-master/WINProgramingProject-master/NGP_Server/NGP_Server/stdafx.h
@@ -8,3 +8,40 @@
 #include <vector>
 #include <string>
 #include<iomanip>
+
+
+#define SERVERPORT 9000
+#define BUFSIZE    512
+using namespace std;
+
+enum EROUND {
+	Round1,
+	MAIN,
+	HELP,
+	Select,
+	Round2,
+	YouDie,
+	YouWin,
+	SelectPlay
+};
+
+typedef struct udpdata
+{
+	char type;
+	char playerID;
+	int playerXpos;
+	int playerYpos;
+	int collisionDamage;
+	EROUND sceneState;
+}udpData;
+
+typedef struct tcpdata
+{
+	char type;
+	char playerID;
+	bool useTeleport;
+	int teleportXpos;
+	int teleportYpos;
+	bool useDash;
+	int hp;
+}tcpData;
